moving everything from bag to bag footer
