+++ src/nft/components/bag/Bag.tsx
@@ -1,9 +1,5 @@
-import { BigNumber } from '@ethersproject/bignumber'
 import { sendAnalyticsEvent } from '@uniswap/analytics'
 import { NFTEventName } from '@uniswap/analytics-events'
-import { useWeb3React } from '@web3-react/core'
-import { GqlRoutingVariant, useGqlRoutingFlag } from 'featureFlags/flags/gqlRouting'
-import { useNftRouteLazyQuery } from 'graphql/data/__generated__/types-and-hooks'
 import { useIsNftDetailsPage, useIsNftPage, useIsNftProfilePage } from 'hooks/useIsNftPage'
 import { BagFooter } from 'nft/components/bag/BagFooter'
 import ListingModal from 'nft/components/bag/profile/ListingModal'
@@ -12,28 +8,10 @@ import { Portal } from 'nft/components/common/Portal'
 import { Column } from 'nft/components/Flex'
 import { Overlay } from 'nft/components/modals/Overlay'
 import { buttonTextMedium, commonButtonStyles } from 'nft/css/common.css'
-import {
-  useBag,
-  useIsMobile,
-  useProfilePageState,
-  useSellAsset,
-  useSendTransaction,
-  useTransactionResponse,
-} from 'nft/hooks'
-import { useTokenInput } from 'nft/hooks/useTokenInput'
-import { fetchRoute } from 'nft/queries'
-import { BagItemStatus, BagStatus, ProfilePageStateType, RouteResponse, TxStateType } from 'nft/types'
-import {
-  buildNftTradeInputFromBagItems,
-  buildSellObject,
-  formatAssetEventProperties,
-  recalculateBagUsingPooledAssets,
-  sortUpdatedAssets,
-} from 'nft/utils'
-import { buildRouteResponse } from 'nft/utils/nftRoute'
-import { combineBuyItemsWithTxRoute } from 'nft/utils/txRoute/combineItemsWithTxRoute'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import { useQueryClient } from 'react-query'
+import { useBag, useIsMobile, useProfilePageState, useSellAsset } from 'nft/hooks'
+import { BagStatus, ProfilePageStateType } from 'nft/types'
+import { formatAssetEventProperties, recalculateBagUsingPooledAssets } from 'nft/utils'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components/macro'
 import { Z_INDEX } from 'theme/zIndex'
 import shallow from 'zustand/shallow'
@@ -103,8 +81,6 @@ const ScrollingIndicator = ({ top, show }: SeparatorProps) => (
 )
 
 const Bag = () => {
-  const { account, provider } = useWeb3React()
-
   const { resetSellAssets, sellAssets } = useSellAsset(
     ({ reset, sellAssets }) => ({
       resetSellAssets: reset,
@@ -118,35 +94,16 @@ const Bag = () => {
     shallow
   )
 
-  const {
-    bagStatus,
-    setBagStatus,
-    didOpenUnavailableAssets,
-    setDidOpenUnavailableAssets,
-    bagIsLocked,
-    setLocked,
-    reset,
-    setItemsInBag,
-    bagExpanded,
-    toggleBag,
-    setBagExpanded,
-  } = useBag((state) => ({ ...state, bagIsLocked: state.isLocked, uncheckedItemsInBag: state.itemsInBag }), shallow)
+  const { bagStatus, bagIsLocked, reset, bagExpanded, toggleBag, setBagExpanded } = useBag(
+    (state) => ({ ...state, bagIsLocked: state.isLocked, uncheckedItemsInBag: state.itemsInBag }),
+    shallow
+  )
   const { uncheckedItemsInBag } = useBag(({ itemsInBag }) => ({ uncheckedItemsInBag: itemsInBag }))
 
   const isProfilePage = useIsNftProfilePage()
   const isDetailsPage = useIsNftDetailsPage()
   const isNFTPage = useIsNftPage()
   const isMobile = useIsMobile()
-  const usingGqlRouting = useGqlRoutingFlag() === GqlRoutingVariant.Enabled
-
-  const sendTransaction = useSendTransaction((state) => state.sendTransaction)
-  const transactionState = useSendTransaction((state) => state.state)
-  const setTransactionState = useSendTransaction((state) => state.setState)
-  const transactionStateRef = useRef(transactionState)
-  const [setTransactionResponse] = useTransactionResponse((state) => [state.setTransactionResponse])
-  const tokenTradeInput = useTokenInput((state) => state.tokenTradeInput)
-
-  const queryClient = useQueryClient()
 
   const itemsInBag = useMemo(() => recalculateBagUsingPooledAssets(uncheckedItemsInBag), [uncheckedItemsInBag])
 
@@ -160,189 +117,13 @@ const Bag = () => {
     }
   }
 
-  const purchaseAssets = async (routingData: RouteResponse, purchasingWithErc20: boolean) => {
-    if (!provider || !routingData) return
-    const purchaseResponse = await sendTransaction(
-      provider?.getSigner(),
-      itemsInBag.filter((item) => item.status !== BagItemStatus.UNAVAILABLE).map((item) => item.asset),
-      routingData,
-      purchasingWithErc20
-    )
-    if (
-      purchaseResponse &&
-      (transactionStateRef.current === TxStateType.Success || transactionStateRef.current === TxStateType.Failed)
-    ) {
-      setLocked(false)
-      setModalIsOpen(false)
-      setTransactionResponse(purchaseResponse)
-      setBagExpanded({ bagExpanded: false })
-      reset()
-    }
-  }
-
   const handleCloseBag = useCallback(() => {
     setBagExpanded({ bagExpanded: false, manualClose: true })
   }, [setBagExpanded])
 
-  const [fetchGqlRoute] = useNftRouteLazyQuery()
-
-  const fetchAssets = async () => {
-    const itemsToBuy = itemsInBag.filter((item) => item.status !== BagItemStatus.UNAVAILABLE).map((item) => item.asset)
-    const ethSellObject = buildSellObject(
-      itemsToBuy
-        .reduce((ethTotal, asset) => ethTotal.add(BigNumber.from(asset.priceInfo.ETHPrice)), BigNumber.from(0))
-        .toString()
-    )
-
-    didOpenUnavailableAssets && setDidOpenUnavailableAssets(false)
-    !bagIsLocked && setLocked(true)
-    setBagStatus(BagStatus.FETCHING_ROUTE)
-    try {
-      if (usingGqlRouting) {
-        fetchGqlRoute({
-          variables: {
-            senderAddress: usingGqlRouting && account ? account : '',
-            nftTrades: usingGqlRouting ? buildNftTradeInputFromBagItems(itemsInBag) : [],
-            tokenTrades: tokenTradeInput ? tokenTradeInput : undefined,
-          },
-          onCompleted: (data) => {
-            if (!data.nftRoute || !data.nftRoute.route) {
-              setBagStatus(BagStatus.ADDING_TO_BAG)
-              setLocked(false)
-              return
-            }
-
-            const purchasingWithErc20 = !!tokenTradeInput
-            const { route, routeResponse } = buildRouteResponse(data.nftRoute, purchasingWithErc20)
-
-            const { hasPriceAdjustment, updatedAssets } = combineBuyItemsWithTxRoute(itemsToBuy, route)
-            const shouldRefetchCalldata = hasPriceAdjustment && purchasingWithErc20
-
-            const fetchedPriceChangedAssets = updatedAssets
-              .filter((asset) => asset.updatedPriceInfo)
-              .sort(sortUpdatedAssets)
-            const fetchedUnavailableAssets = updatedAssets.filter((asset) => asset.isUnavailable)
-            const fetchedUnchangedAssets = updatedAssets.filter(
-              (asset) => !asset.updatedPriceInfo && !asset.isUnavailable
-            )
-            const hasReviewedAssets = fetchedUnchangedAssets.length > 0
-            const hasAssetsInReview = fetchedPriceChangedAssets.length > 0
-            const hasUnavailableAssets = fetchedUnavailableAssets.length > 0
-            const hasAssets = hasReviewedAssets || hasAssetsInReview || hasUnavailableAssets
-            const shouldReview = hasAssetsInReview || hasUnavailableAssets
-
-            setItemsInBag([
-              ...fetchedUnavailableAssets.map((unavailableAsset) => ({
-                asset: unavailableAsset,
-                status: BagItemStatus.UNAVAILABLE,
-              })),
-              ...fetchedPriceChangedAssets.map((changedAsset) => ({
-                asset: changedAsset,
-                status: BagItemStatus.REVIEWING_PRICE_CHANGE,
-              })),
-              ...fetchedUnchangedAssets.map((unchangedAsset) => ({
-                asset: unchangedAsset,
-                status: BagItemStatus.REVIEWED,
-              })),
-            ])
-
-            let shouldLock = false
-
-            if (hasAssets) {
-              if (!shouldReview) {
-                if (shouldRefetchCalldata) {
-                  setBagStatus(BagStatus.CONFIRM_QUOTE)
-                } else {
-                  purchaseAssets(routeResponse, purchasingWithErc20)
-                  setBagStatus(BagStatus.CONFIRMING_IN_WALLET)
-                  shouldLock = true
-                }
-              } else if (!hasAssetsInReview) setBagStatus(BagStatus.CONFIRM_REVIEW)
-              else {
-                setBagStatus(BagStatus.IN_REVIEW)
-              }
-            } else {
-              setBagStatus(BagStatus.ADDING_TO_BAG)
-            }
-
-            setLocked(shouldLock)
-          },
-        })
-      } else {
-        const routeData = await queryClient.fetchQuery(['assetsRoute', ethSellObject, itemsToBuy, account], () =>
-          fetchRoute({
-            toSell: [ethSellObject],
-            toBuy: itemsToBuy,
-            senderAddress: account ?? '',
-          })
-        )
-
-        const { updatedAssets } = combineBuyItemsWithTxRoute(itemsToBuy, routeData.route)
-
-        const fetchedPriceChangedAssets = updatedAssets
-          .filter((asset) => asset.updatedPriceInfo)
-          .sort(sortUpdatedAssets)
-        const fetchedUnavailableAssets = updatedAssets.filter((asset) => asset.isUnavailable)
-        const fetchedUnchangedAssets = updatedAssets.filter((asset) => !asset.updatedPriceInfo && !asset.isUnavailable)
-        const hasReviewedAssets = fetchedUnchangedAssets.length > 0
-        const hasAssetsInReview = fetchedPriceChangedAssets.length > 0
-        const hasUnavailableAssets = fetchedUnavailableAssets.length > 0
-        const hasAssets = hasReviewedAssets || hasAssetsInReview || hasUnavailableAssets
-        const shouldReview = hasAssetsInReview || hasUnavailableAssets
-
-        setItemsInBag([
-          ...fetchedUnavailableAssets.map((unavailableAsset) => ({
-            asset: unavailableAsset,
-            status: BagItemStatus.UNAVAILABLE,
-          })),
-          ...fetchedPriceChangedAssets.map((changedAsset) => ({
-            asset: changedAsset,
-            status: BagItemStatus.REVIEWING_PRICE_CHANGE,
-          })),
-          ...fetchedUnchangedAssets.map((unchangedAsset) => ({
-            asset: unchangedAsset,
-            status: BagItemStatus.REVIEWED,
-          })),
-        ])
-        setLocked(false)
-
-        if (hasAssets) {
-          if (!shouldReview) {
-            purchaseAssets(routeData, false)
-            setBagStatus(BagStatus.CONFIRMING_IN_WALLET)
-          } else if (!hasAssetsInReview) setBagStatus(BagStatus.CONFIRM_REVIEW)
-          else {
-            setBagStatus(BagStatus.IN_REVIEW)
-          }
-        } else {
-          setBagStatus(BagStatus.ADDING_TO_BAG)
-        }
-      }
-    } catch (error) {
-      setBagStatus(BagStatus.ADDING_TO_BAG)
-    }
-  }
-
-  useEffect(() => {
-    useSendTransaction.subscribe((state) => (transactionStateRef.current = state.state))
-  }, [])
-
   useEffect(() => {
     if (bagIsLocked && !isModalOpen) setModalIsOpen(true)
   }, [bagIsLocked, isModalOpen])
-
-  useEffect(() => {
-    if (transactionStateRef.current === TxStateType.Confirming) setBagStatus(BagStatus.PROCESSING_TRANSACTION)
-    if (transactionStateRef.current === TxStateType.Denied || transactionStateRef.current === TxStateType.Invalid) {
-      if (transactionStateRef.current === TxStateType.Invalid) setBagStatus(BagStatus.WARNING)
-      else setBagStatus(BagStatus.CONFIRM_REVIEW)
-      setTransactionState(TxStateType.New)
-
-      setLocked(false)
-      setModalIsOpen(false)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [transactionStateRef.current])
 
   const hasAssetsToShow = itemsInBag.length > 0
 
@@ -389,7 +170,7 @@ const Bag = () => {
               {isProfilePage ? <ProfileBagContent /> : <BagContent />}
             </Column>
             {hasAssetsToShow && !isProfilePage && (
-              <BagFooter fetchAssets={fetchAssets} eventProperties={eventProperties} />
+              <BagFooter setModalIsOpen={setModalIsOpen} eventProperties={eventProperties} />
             )}
             {isSellingAssets && isProfilePage && (
               <Box
